Change ISDIMM PN to use FN
There was a bug discovered on Bonnell where the PN was not being
reported the same as the FN which is not spec compliant. To correct this
a targeted change was made to set the PN the same as the FN after all
the calculations have occurred to determine the PN and various other
data from the SPD since most of the data is reliant on the PN to derive.

Change-Id: I702f2f4a99cfbc3a34eb7b2af8ae8b7fc0124346
EWM: 598195
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/166100
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>