Ody Flash Check: Additional OCMB_FLASH_ERROR FSM Paths
This commit adds two new paths for OCMB_FLASH_ERROR
to the Ody code update FSM. When a flash error happens on
SIDE 0 after a golden boot (and forced code update) was performed,
the FSM will switch the OCMB to SIDE 1 and attempt to
perform check for ready. If the flash error happens again
on SIDE 1 (after gold side boot/code update), then the OCMB
will be deconfigured and garded.

Change-Id: I821ffb0945032b82a7afd8693111b307b77c124f
STGD: 601664
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/166957
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>