PRD: Update ScomRegister class to differentiate with signature offset
As is, PRD's ScomRegister class when stored checks the address
and target type values to determine uniqueness of the object.
However with Odyssey/Explorer OCMBs, there are now registers
that share both address and target type. Therefore this commit
adds in checking the signature offset defined in the rule code as
well when determining the uniqueness of the ScomRegister object.

Change-Id: I1fb59afcf4f327de456ef4d577f37f5276e03caa
STGD: 599069
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/prd/+/166977
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Austin F Cui <austin.cui@ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/167662
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>