Terminate System if both "Disable TPM" and "TPM Required" are set
This commit checks to see if both the "Disable TPM" and the
"TPM Required" system policies are set.  If they both are, an
error log is created and the system immediately shuts down.
The error log message instructs the user to only set one of
these values at a time.

Change-Id: I76034afe05559d418e80245cf592a0c09c880a99
STGD: 589636
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/167245
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Chris Engel <cjengel@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: MIKE VAGELINE <mpvageli@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>