Avoid eecache accesses in DDIMM SPD parking
After every i2c operation to the SPD we park the eeprom pointer
at a safe address to lessen the potential for data corruption
due to random i2c signals coming from the processor.  During
initial presence detection if we are populating the eecache for
that part we won't yet have a valid cache entry.  That means if
we try to access data from the cache (which is the default for
most generic spd/vpd operations) we will get an error.  That
error has the side-effect of disabled ALL cache access until
the next IPL.

Change-Id: Ie5ec990e18cb50f181f7c20df7ee2d5bfdc9b6f9
STGD: 599359
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/167382
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>