Enhancements for patch-bmc
Added the following enhancements:

1. Added a new option, PATCH_BMC_NO_DEBUG. This allows the customization
of whether or not to copy over to the BMC the large debug binaries.
This helps the efficiency of test and debug when the debug binaries
are not needed (speeds up the transfer times).

2. Added a validation check to assure the BMC is HostState.Off.  This
helps eliminate the inadvertant application of patches when the Host
is on and the pldmd refresh would not be appropriately done in the
proper expected state.

3. Added a sleep delay to make sure the rsync is successful before
removing the directory.

Change-Id: I1561149bc4808335b541624156e7d6a9b597801a
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/170705
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>