Odyssey code update: Always check async FFDC bit during SPPE boot
This commit changes Hostboot's behavior when an async error occurs
during boot.

Previously, if the SPPE was "not ready" according to the
check_for_ready HWP, Hostboot would assume that it was not reachable
via chipops, and it would not retrieve async FFDC even if the bit is
set. This behavior is a bit too conservative, because the SPPE may be
"not ready" because it did not reach runtime, but still retain the
ability to respond to chipops, and if the async FFDC bit is set, we
want to retrieve the FFDC from the SPPE if at all possible for error
diagnostics.

This commit makes a change so that even if the SPPE does not reach
runtime, we check the async FFDC bit, and if it is set, we try to
retrieve the FFDC from the SPPE.

There may occur cases when the SPPE sets the async FFDC bit, does not
reach runtime, and cannot respond to chipops. In this case, we will
get a chipop timeout and treat the SPPE as dead.

Change-Id: I1dfb43208f44d849158486d72260ef2435919f27
STGD: 600384
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/167761
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>