Odyssey code update: Perform forced sync if normal sync fails
This commit adds a forced code level sync after failed normal syncs to
the Odyssey code update FSM action executor.

This fixes a defect where a normal image sync may fail on the SBE
because the hash check code (due to corruption or for any reason), and
Hostboot takes this as a prompt to deconfigure the OCMB, rather than
trying again with a forced sync which ignores the hash check and might
succeed.

If the forced sync fails after the normal sync fails, we still
deconfigure the chip.

Change-Id: I5864adf644c96763c7f2b98d25486d491f136deb
STGD: 597383
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/167777
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>