Update Secure Version to 2 for Hostboot Standalone
This commit makes the default secure version equal to 2 when building
images for hostboot standalone.  This will be especially helpful
when the setup-standalone-repo and patch-bmc scripts are used to
patch code on both HW or simics systems with security enabled because
currently all formal builds (both development/lab/imprint and
production) are built with a secure version of 2.

Change-Id: I085cbc7d1145185af54836a9c687d3d6d15dee73
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/168466
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>