Odyssey code update: Handle OCMBFW PNOR changes mid-OCMB-update
Previously, there was an issue where if host firmware (particularly
the OCMBFW partition) is updated after the Odyssey firmware from the
old firmware is flashed onto an OCMB, then the Odyssey code update
system will detect that the firmware written to the OCMB doesn't match
what's in the (updated) PNOR, and presume that the code update failed.

This commit fixes the issue by keeping a hash of all the firmware
related to a given Odyssey OCMB in an attribute on that OCMB, and then
recalculating that hash on each boot and comparing with the attribute
to see whether PNOR has changed since the previous boot. If the
firmware has changed, the FSM state is reset, which restarts the
boot/update process from the beginning.

(This works even when a firmware update is not in progress, because
the FSM state is reset on every fresh boot anyway.)

Change-Id: I4078e991a8b72bea068d8cd761ff370a10e21cbc
STGD: 596733
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/168495
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>