DDR5 2U Telemetry Handling
Fix the DDR5 2U Telemetry and Health Check handling.

The isDDR5_Health_Check logic is to aid the determination of
whether the invocation is to process the DDR5 4U Health Check logic.

When refactoring the original logic, the processing of the DDR5 2U
Telemetry logic flow was inadvertantly omitted from special handling.

When DDR5 Health Check callback is invoked, special handling needs
to catch the DDR5 2U cases (isDDR5_Health_Check will evaluate to
not run the DDR5 Health Check, however the cases in the main flow
need to conditionally catch and qualify the DDR5 2U as Telemetry
only case).

STGD: 605387
Change-Id: I5d65bd1fc9a9c96e527aa44fa0146072d85de1e7
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/168517
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>