Removed sbeAttrSync tag from non Odyssey SBE attributes
If an attribute is marked for syncing from Hotsboot/Cronus to Odyssey
SBE ie.toSBE="1", and if that attribute is not consumed by Odyssey SBE
(not present in Odyssey SBE filter XML), then SBE would send attribute not
found error for those attributes. If there are many such attributes, then
it would unnecessarily increase the size of the attribute update blob
which increases the time taken for the attribute sync istep.

If both toSBE="0" and fromSBE="0", then the entire sbeAttrSync tag can be
removed.

SBE attrtool has been updated to catch such errors in CI. That tool changes
will be rebased on top of this commit's SBE mirror commit, for the SBE CI
to pass.

Change-Id: Idc97ca8a0e6eb8f80282c7ef0b0a98115b04f505
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/ekb/+/168576
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Kahn C Evans <kahnevan@us.ibm.com>
Reviewed-by: Louis Stermole <stermole@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/168579
Tested-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>