Avoid i2c contention during odyssey sbe dumps
The BMC/FSP dump code will use i2c to grab the data.  There is no
handshake between that code and Hostboot.  Since Hostboot has other
threads running that may attempt i2c operations on the same engine
that is used for the dump, we need to block those to avoid any
multimaster contention issues.  This is done by locking the
mutex that we already use to handle our internal threads.

Change-Id: I9a5f20ec1db45371b9a362ce54e04a75533bc782
STGD:605585
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/168583
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>