Use distinct user details section numbers for raw sbe data
We don't currently parse some of the binary data that we log
for SBE errors so it displays as raw hex.  However, we could
enhance the parser in the future to support this data and/or
other offline parsers could be written to support them.  To
better facilitate this potential future enhancements we should
create the data with unique section ids that can be keyed on.

Change-Id: I2b0cd517073b67ec3da21826791976a45890c63e
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/169144
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>