Force previous errors to be recovered before code update
The fact that we're doing a code update means that any errors we
hit prior to that are from a downlevel version. Therefore none of
those logs should show up as visible errors.

From the Odyssey Code Update and Error Handling flow, this would
be the following path:
 ody_sppe_check_for_ready = Fail
 RC_POZ_SPPE_NOT_READY_ERR? = Yes
 Async FFDC? = Yes
 Is active side current? = No
 Was the active side updated before? = No
 Commit RC as Recovered <<< this action was missed sometimes

This bug was discovered when this scenario hit an error doing
the initial code update operation itself.  Those errors were
properly downgraded as we moved to golden, but the original
async logs were not.  That led to visible logs with callouts
and also deconfiguration which prevented the code update from
fixing the issue as it should have.

Change-Id: I3b7fd1dd285e09c9205ace701dcd6ca6cb23425c
STGD: 607039
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/169653
Tested-by: Mike Baiocchi <baiocchi@ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>