Fix mirror glitch with WOF disable change
A change last year incorrectly mirrored, which this corrects.

This will not have WOF disabled if the the WOF table override
checking actually failed if we’re in
ATTR_SYSTEM_WOF_VALIDATION_MODE==WARN | INFO modes.  The present
default for the validation attribute is INFO so that says we’ll
log an INFO error and just truck ahead with WOF on even through
the overrides got applied in some skewed way.

The correct code was already released in FW1050 so this just
restores the same behavior that is already in the field.

Change-Id: I99a8e57b9d94e4b37bfcdacf0c69ed42a4b00a42
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/170367
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>