Enable patch-bmc to allow empty rsync sources
Wildcard substitution with rsync when source directory is empty fails
with rsync: link_stat failure due to no such file or directory.

rsync will properly skip the empty directory when specifying the path with
a proper / at the end of the directory name.

Cleanup the HBRT_PATCH_DIR when completed the processing.

Change-Id: I7047add8f6ab7960cfe3b12148e051c01a2d8480
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/170321
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Sheldon Bailey <baileysh@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>