Added RC_POZ_STREAM_GET_FAILED
sbefifo_hwp_data_istream::get() method has been modified to return
ReturnCode to the caller. Hence, in case of failure, FAPI RC must be
created and returned to the caller. This RC has been added to create
FFDC if the caller tries to read when words left is zero or data
pointer is NULL.

EWM: 611872
Change-Id: I66df3dd146b4dbceb6a55fc62cbca16af35ac8b9
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/ekb/+/170450
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Mayur Bansal <mayur.bansal1@ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/170475
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>