Fix SBE FFDC FSP parser
There was an issue in the FSP SBE FFDC parser where it expected the RC
to be a part of the blob of data sent to it for the Odyssey and P10 proc
parsers. Hostboot does not send this data to its parser so that logic
was removed and a hardcoded RC is put in its place.

Change-Id: I7506376eca29024c329b9a5d4ca58bb356cdf6fc
STGD:611870
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/170838
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>