Handle crashes in spi flash failures by side switching
There was an issue where the SBE could crash during the spi flash check
chip-op which is to be addressed in 604644. To better handle these
scenarios in Hostboot, the FSM will now switch sides in the event of a
crash eventually leading to a code update from the golden side if
necessary.

Change-Id: I42f50302fa2325c29d7cb3e126969855fbe03905
STGD:609300
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/170839
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>