Add calls to sendProgressCode() during Odyssey Code Update
On fully loaded memory systems, the time it takes to pass all of the
Odyssey code updates, especially if the updates are occurring over I2C
in istep 11.4, takes too long and BMC's watchdog timer is triggered.
Assuming that Hostboot has hung, the BMC brings the system down.

This commit fixes the problem by sending down a progress code to the BMC
via INITSERVICE::sendProgressCode() in key parts of the code update
path.  This action resets the watchdog timer and allows all of the
code updates to complete successfully.

Change-Id: If765f3cba3663a993c9945014849a9a9c340216e
STGD: 613644
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/171107
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>