Do not fail istep for recovered errors from chipop
HWPs may log recoverable RCs but return success.  However,
all RCs are returned from the chipop and then processed into
error logs that are returned to the caller.  If the caller
assumes that any returned log is a failure that will lead to
the istep failing.

The change here is to only fail the HWP chipop if there is
actually bad status returned from the chipop or it fails
in some other way.

A new errlentry interface is added to do the above check:
  hasErrorTypeOtherThan(errortype)
This will return true if there is a single log inside the
errlog aggregation that has an error type other than what
is passed in (inverse of hasErrorType).

Change-Id: Ic466c67a9f3426640cc35b8cb03ecebe7890303e
STGD: 613513
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/171339
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>