Hostboot Runtime Odyssey SBE Dump
Add the PLDM based Odyssey SBE dump call for the Odyssey recovery error path.

This will enhance the IPL Odyssey dump flows to perform the relevant
equivalents during Hostboot Runtime.

During Hostboot Runtime there are limitations on how much of the PLDM
semantics are available for implementation, thus a simple timeout
is implemented in the rare cases of requiring a dump of the Odyssey
at HBRT.

Added an rt_cmd dumpSbe PHYP command line interface which
facilitates the testing and collecting of an Odyssey dump, however,
the test mechanism is not the actual logic flow used during the
normal odyssey_recovery_handler flow, so be aware of the differences
if the rt_cmd dumpSbe is used for testing purposes.

JIRA: PFHB-733
STGD: 613499
Change-Id: Ie4fb395fb9cb94bb2b2e93c3bfce7629358bb3d6
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/171432
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/171960