Remove full image spiflash check from manufacturing mode
There is a batch of MSL DDIMMs that were built with bad ECC in
the unused portion of the Odyssey SBE flash.  These parts will
fail if we run the full check and there is no way to fix them.
This check will happen instead during dimm build so it does not
need to be part of the firmware boot.

Change-Id: Ibeb5643a62b2a2ffa7c6ea564e002391c42539df
STGD: 615647
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/171411
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>