Do not attempt recovery for spi flash errors in mfg mode
In the field we will attempt a variety of recovery options
(side switches, code updates) to get around ECC errors in the
odyssey sbe spi flash.  However, in manufacturing we want to
throw an error so the suspect part can be replaced.  The only
recovery we want is for a downlevel code level in case there
is a fix to the spiflash chipop logic itself.

Change-Id: I386af7fb1af252a8417477d160eea462bae30970
STGD: 617246
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/171872
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>