Odyssey HWP Error Handling
Some of the Odyssey HWP errors could be attributed to a cause
which is signified by the FIR bits. The PRD code already
parses these bits. There is a new HWP that needs to be called
by Hostboot logic when there is an Odyssey HWP error called
ody_blame_firs() to know if the fir bits are set or not. If
the bits are set, leverage the PRD logic to help interpret
those bits. Also, implemented some changes to the logic to
handle some of the error path scenarios.

This commit also included delayed deconfig actions to ody_upd_fsm.
Basically added delayed_deconfigure_ocmb and
delayed_deconfig_gard_ocmb actions to ody_upd_fsm.
Previously, only deconfigure_ocmb and deconfig_gard_ocmb
actions were supported. Delayed deconfig allows PRD to run and check
for attentions before the deconfig happens at the end of the istep.

JIRA: PFHB-518
STGD: 613473
Change-Id: I844633044403adcc3da005d0765e9916080ac0d2
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/172249
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>