HTMGT: Move PREHEAT_PERCENT from DIMM to OCMB
Originally the PREHEAT_PERCENT was an attribute on the DIMM, but this
commit will move it to the OCMB. (under chip-obmc-generic so it works
under Bonnell as well)

The PREHEAT_PERCENT attribute on an OCMB target will be used for each
of the DIMMs under that OCMB.

HTMGT will now be sending version 2 of the Memory Power config packet
which will contain the full power (in addition to the preheat power)
under the OCMB.

HTMGT will now use 0 for the minimum utilization for preheat, so the
default points will be: 0%, 25%, 50%, 75%, and 100%

HTMGT will now log an error 2628 error if the burst length or clock
attributes are not valid.

Commit also updates P11 OCC lid for simics.

Change-Id: I2af7d8cb06b1cc20be6109d95510bb6e3221dfab
JIRA: PFES-73
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/172270
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Sheldon Bailey <baileysh@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>