Ignore PMIC Health Check callbacks for DDR4
The new PMIC Health Check calls (which are for DDR5) are not performed
for DDR4 (DDR4 4U only runs the Telemetry HWPs); however the timer
which is used to trigger the periodic health check callbacks needs to be
filtered and only applied to DDR5 systems.

The terminology of the internals of the PMIC Telemetry and PMIC Health Check
calls has a very inconsistent naming convention with legacy methods which makes
some of the external tracing to be confusing, so cleaning up some of the trace
points to help alleviate any external confusion.

The determination of PMICs being either DDR4 or DDR5 is determined during
the invocations of each of the Telemetry and Health Check methods.  This
makes the calls for the Telemetry and Health Check methods required in order
to identify the type and then the execution of the applicable methods.

STGD: 620612
Change-Id: Idacae0c22b0f0627a4757bd0523246de33cbeedf
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/172731
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>