Enable default hexdump for hwpf User Detail parsing
Add a check to see if the subType being requested is defined
in the hwpfUserDetailDataSubSection dictionary.  If a new
subType is being requested, process the incoming new subType
as a default hexdump so no python exceptions will be surfaced.

This component is likely to have new subTypes more frequently
than some of the other parsers, so enabling this default
hexdump will allow the BMC peltool to not throw python
exceptions to the end user.

If python exceptions are thrown by the peltool, the
parsing output will default also to hexdump the data, however
the end user will see python exceptions at the command line
which is not desired.

STGD: 621570
Change-Id: I98c6b5a01cc4b1ea55775f0ba02cd48beb6490d9
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/173314
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>