Enforce buffer limit for getLidFromBMC
Use cases which flow via the getLidFromBMC need to enforce a buffer
size check to prevent data overflow when the caller has not allocated
the proper size buffer.

This use case occurs when the payload size, i.e. lid content, is not known
prior to the invocation of the getLidFromBMC.  The caller however needs
to pre-allocate the memory for which the payload will eventually be copied.

In other workflows, the getLidFileFromOffset uses will wrap other
logic around either the EOF or the PLDM completion code of
PLDM_DATA_OUT_OF_RANGE.

Change-Id: Ic507c37d9573f3dbe834e19d5d977fe32bdb790e
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/176969
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>