Enable dumpSbe at Hostboot Runtime
During the development and test of the Hostboot Runtime component
of PLDM::dumpSbe invocation, the ifdef was not updated to allow the
calling of the dumpSbe method during Hostboot Runtime.

The unit test of the dumpSbe method was accomplished by using
the rt_cmd entry points and customizing the hard failure flows.

End-to-end testing was not performed at that time which is why
this was missed during implementation.

STGD: 624322
Change-Id: I63fb42b24ff2c07fbf57c62f71afbfe28bb60354
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/173341
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>