Add algorithm for spare core selection
Adds the logic necessary to choose an arbitrary number of spare CORE/FCs
to hwas common code. This logic is executed after FCO as it needs to
know the state of things before FCO deconfigs are applied in case a
CORE/FC needs to be reclaimed from FCO to be considered spare.

Adds logic to standalone simics startup that creates spare core
entries in proc1.  A testcase was added to verify the results.

Fixes some mismatched logic related to spare core counting that
did not agree on CORE vs FC counting.

Includes updates to our prebuilt EECACHE files to reflect the new
spare core bits.

Change-Id: Ifbf9b042a921d2a88bf5815f7c72d7c33cc54f47
JIRA: PFHB-777
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/173751
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Shawn Hanson <shanson2@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>