Support Offline PEL Parsing tooling
Provide a script that will properly setup and configure the
offline pel parsing environment.

The script will support the Jenkins invocation as well as
a copy of the script within a hostboot workon or independently
on a cloud lcb with only proper networking to establish
download of the peltool-wrapper and associated packages.

At a future time the Jenkins job will be updated with
a clone of the hostboot repo which will contain this
commit and then be able to maintain the source in a single
repository.

Jenkins job -> Offline_Peltool_Parsing

Change-Id: I319ac784949135df2600e9b98513c1e6ab9231fa
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/173753
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Connor Leu <connor.leu@ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Veena Ganti <veena@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/176513