Adjust string trace hashing method
With the inclusion of Odyssey, we've run into some
string hash collisions. Due to the nature of our trace hash
technique, there is no guaranteed or easy way to "fix"
trace collisions or ensure they do not occur. As such,
we are deciding to fail the build before generating the
.elf and .bin files, and dump out any hash collisions so
that they can be manually fixed by a developer.

JIRA: PFHB-480
Change-Id: Ifecd6a66d50ebcb7eba2fffdd265b8d4e9ccbdbe
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/173825
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>