Update eBMC trace plugin to handle OCC trace data from HTMGT
HTMGT will add OCC traces to some PELs with a Sub-section type of 2.
That data was not being parsed by the b2a00.py plugin.

Also updated hostfw_trace.py to handle truncated trace blocks.
This handles when the offset after last trace entry is larger then the
trace data provided. When that offset is larger, it will get set to the
endOffset from the trace header.

Change-Id: Ib19b6d1a29b58db63c364b36742635de116fc1fd
STGD: 620150
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/174361
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>