Double fapi delay on Denali
Hostboot has an incorrect timebase calculation on Denali because
it runs at a higher core floor frequency and our logic needs to
know the answer before we have the ability to determine that.
As a workaround for that issue we will increase the value for
fapi2::delay by a factor large enough to compensate for the
incorrect timebase.  This should guarantee that any required
hardware delays are long enough.

STGD: 626449
Change-Id: I22b6e098d72c377004d743a24b298d82a164bf2b
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/173947
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>