Force FAPI2_RC_PLAT_ERR_SEE_DATA package to info severity
The SBE returns an invalid severity for the FFDC package with
RC=FAPI2_RC_PLAT_ERR_SEE_DATA.  We convert an invalid value
to Unrecoverable because we treat it as a code bug.  In this
case we know that the package contains no error/callout info
so we are going to just force it to informational.

Change-Id: I3fcbc55815162b927530d9b9852f09901c37c2ac
STGD:628020
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/174143
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Connor Leu <connor.leu@ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>