Delay initial PMIC telemetry calls by 1 hour
The PMIC telemetry data collection seems to take a variable amount
of time which can add add up to a significant amount on a large
system.  Lab observations have shown the range to be anywhere from
a couple minutes to half an hour for a fully populated system.
Since HBRT is serialized, no other work can be done while this is
happening so useful work (PM complex enablement) gets delayed.

Change-Id: If8888c81bff5839922d366e5fdf6031f6910d0c4
STGD: 630397
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/174708
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ellen M Bauman <emb@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>